Added support for context-local request and response objects.
Some objects used by bottle need to be context local. By default, these
are local to the current thread: The instance’s values are different for
each separate threads. In environments where 'light' or 'mirco' threads
are used to parallelize requests, a singel thread may contain several
contexts and thread-locality is not sufficiant anymore. The new function
:func:`set_context_ident` is used to change the callable that is supposed
to return a context identifying object when called from within a context.
It defaults to thread.get_ident.