StplTemplate speedup. STPL now beats all other template engines in the jinja2 benchmark (generating a 1000x10 cell table).
Stpl used to automatically encode byte-strings returned by inline
statements and needed to check the return type to do so. No other template
engine does that.

Inline-statements that do not return unicode are limited to ASCII now. They
are not automatically encoded anymore. Use unicode or encode manually if you
need to generate non-ASCII characters. In return, rendering speed is much
better if you use a lot of {{inline-statements}}.